requirements.txt
PyQt5>=5.15
pyyaml
numpy
opencv-python
anomalib>=2.2.0 ; python_version>='3.10'
torch>=2.1
imagingcontrol4